backend/requirements.txt
Flask
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-CORS
Flask-Restful
redis
PyJWT
